Move pricing services back into pricing module
