+++ src/telliot/pricing/bittrex.py
