--- src/telliot/reporter_plugins/rinkeby_btc_usd/datafeed_utils/__init__.py
@@ -1,1 +1,0 @@
-"""BTC price getters."""
