+++ src/telliot/reporter_plugins/rinkeby_btc_usd/registry.py
@@ -3,16 +3,16 @@ import statistics
 
 from telliot.datafeed.asset_price_feed import AssetPriceFeed
 from telliot.datafeed.asset_price_source import AssetPriceSource
-from telliot.reporter_plugins.rinkeby_btc_usd.datafeed_utils.bittrex import (
+from telliot.pricing.bittrex import (
     BittrexPriceService,
 )
-from telliot.reporter_plugins.rinkeby_btc_usd.datafeed_utils.coinbase import (
+from telliot.pricing.coinbase import (
     CoinbasePriceService,
 )
-from telliot.reporter_plugins.rinkeby_btc_usd.datafeed_utils.coingecko import (
+from telliot.pricing.coingecko import (
     CoinGeckoPriceService,
 )
-from telliot.reporter_plugins.rinkeby_btc_usd.datafeed_utils.gemini import (
+from telliot.pricing.gemini import (
     GeminiPriceService,
 )
